Rename currencyList KT-4
