+++ src/store/reducers/converter.ts
@@ -1,5 +1,5 @@
 import {StorageService} from 'api/service';
-import {CONVERSION_FORM_ONE, CONVERSION_FORM_TWO} from 'core/constants/currencyList';
+import {CONVERSION_FORM_ONE, CONVERSION_FORM_TWO} from 'core/constants/constants';
 
 import {ConverterAction, ConverterActionTypes, IFormState} from 'store/actions/converter';
 import {RootState} from '.';
